requirements.txt
anthropic
assemblyai
cryptography
flask
flask-login
gunicorn
openpyxl
psycopg2-binary
python-dotenv
requests
werkzeug
